requirements.txt
anthropic==0.3.7
python-dotenv==0.19.0
Flask==2.0.3
Werkzeug==2.0.3
gunicorn==20.1.0